+++ CHANGES.md
@@ -1,5 +1,10 @@
 # **CLASP.Ruby** Changes
 
+## 0.18.0 - 10th April 2019
+
+* ~ changed *Alias classes to *Specification
+* ~ CLASP:Arguments#aliases attribute is now changed to #specifications, and a [DEPRECATED] #aliases added
+
 ## 0.17.0 - 10th April 2019
 
 * + added CLASP::Arguments.load(), which allows to load argument-specifications from Hash or from YAML